# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Papafi Backend API",
    "version": "1.0.0",
    "description": "Papafi wallet backend API for frontend and mobile clients. Documents implemented Express routes only."
  },
  "servers": [
    { "url": "https://api-staging.papifi.com" },
    { "url": "https://api.papifi.com" },
    { "url": "http://localhost:3000" }
  ],
...
  "paths": {
...
    "/api/wallet/create/{userId}": {
      "post": {
        "tags": ["Wallet"],
        "summary": "Create NGN wallet and virtual account",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/UserIdPath" }],
        "responses": {
          "201": { "description": "Wallet created.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WalletResponse" } } } },
          "200": { "description": "Existing wallet returned.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WalletResponse" } } } },
          "403": { "$ref": "#/components/responses/OkFalseError" }
        }
      }
    },
    "/api/wallet/create-usd/{userId}": {
      "post": {
        "tags": ["Wallet"],
        "summary": "Request USD virtual account",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/UserIdPath" }],
        "responses": {
          "201": { "description": "USD account request returned from provider.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UsdAccountRequestResponse" } } } },
          "400": { "$ref": "#/components/responses/OkFalseError" },
          "403": { "$ref": "#/components/responses/OkFalseError" }
        }
      }
    },
    "/api/wallet/balance/{userId}": {
      "get": {
        "tags": ["Wallet"],
        "summary": "List authenticated user's wallets and balances",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/UserIdPath" }],
        "responses": {
          "200": { "description": "Wallets returned.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WalletListResponse" } } } },
          "403": { "$ref": "#/components/responses/OkFalseError" }
        }
      }
    },
    "/api/wallet/withdraw": {
      "post": {
        "tags": ["Wallet"],
        "summary": "Withdraw from wallet to bank account",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WithdrawRequest" } } }
        },
        "responses": {
          "200": { "description": "Withdrawal submitted or duplicate returned.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransactionProviderResponse" } } } },
          "400": { "$ref": "#/components/responses/OkFalseError" },
          "403": { "$ref": "#/components/responses/OkFalseError" },
          "502": { "$ref": "#/components/responses/OkFalseError" }
        }
      }
    },
    "/api/wallet/cards/create": {
      "post": {
        "tags": ["Cards"],
        "summary": "Create a virtual card",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateCardRequest" } } }
        },
        "responses": {
          "200": { "description": "Safe card metadata. PAN/CVV are not returned.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CardResponse" } } } },
          "400": { "$ref": "#/components/responses/OkFalseError" },
          "404": { "$ref": "#/components/responses/OkFalseError" }
        }
      }
    },
    "/api/wallet/cards/{id}/fund": {
      "post": {
        "tags": ["Cards"],
        "summary": "Fund a virtual card from wallet balance",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/CardIdPath" }, { "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CardMoneyMovementRequest" } } }
        },
        "responses": {
          "200": { "description": "Card funding submitted or duplicate returned.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransactionProviderResponse" } } } },
          "400": { "$ref": "#/components/responses/OkFalseError" },
          "403": { "$ref": "#/components/responses/OkFalseError" }
        }
      }
    },
    "/api/wallet/cards/{id}/withdraw": {
      "post": {
        "tags": ["Cards"],
        "summary": "Withdraw funds from card to wallet",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/CardIdPath" }, { "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CardMoneyMovementRequest" } } }
        },
        "responses": {
          "200": { "description": "Wallet credited from card withdrawal.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransactionProviderResponse" } } } },
          "400": { "$ref": "#/components/responses/OkFalseError" },
          "403": { "$ref": "#/components/responses/OkFalseError" }
        }
      }
    },
    "/api/wallet/cards/{id}/freeze": {
      "post": {
        "tags": ["Cards"],
        "summary": "Freeze an owned virtual card",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/CardIdPath" }],
        "responses": {
          "200": { "description": "Card frozen.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CardMessageResponse" } } } },
          "404": { "$ref": "#/components/responses/OkFalseError" }
        }
      }
    },
    "/api/wallet/cards/{id}/unfreeze": {
      "post": {
        "tags": ["Cards"],
        "summary": "Unfreeze an owned virtual card",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/CardIdPath" }],
        "responses": {
          "200": { "description": "Card unfrozen.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CardMessageResponse" } } } },
          "404": { "$ref": "#/components/responses/OkFalseError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Required for money movement endpoints. Must be a stable unique string of at least 8 characters per client operation. The backend also accepts idempotencyKey in JSON body.",
        "schema": { "type": "string", "minLength": 8 },
        "example": "wd_01hzz6n5qx9m8v2k"
      },
...
      "UserIdPath": { "name": "userId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
      "CardIdPath": { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
...
    },
    "responses": {
...
      "OkFalseError": { "description": "Error response.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OkFalseError" }, "example": { "ok": false, "message": "Invalid transaction PIN" } } } },
...
    },
    "schemas": {
      "Currency": { "type": "string", "enum": ["NGN", "USD", "GBP"] },
...
      "OkFalseError": { "type": "object", "properties": { "ok": { "type": "boolean", "example": false }, "message": { "type": "string" } }, "required": ["ok", "message"] },
...
      "Wallet": { "type": "object", "properties": { "id": { "type": "string", "format": "uuid" }, "mapleradAccountId": { "type": "string", "nullable": true }, "accountNumber": { "type": "string", "nullable": true }, "bankName": { "type": "string", "nullable": true }, "currency": { "$ref": "#/components/schemas/Currency" }, "balance": { "type": "string", "example": "0.00" }, "availableBalance": { "type": "string", "example": "0.00" }, "pendingBalance": { "type": "string", "example": "0.00" }, "ledgerBalance": { "type": "string", "example": "0.00" }, "usdAccountId": { "type": "string", "nullable": true }, "usdAccountStatus": { "type": "string", "enum": ["pending", "approved", "rejected"] }, "createdAt": { "type": "string", "format": "date-time" }, "updatedAt": { "type": "string", "format": "date-time" } } },
      "WalletResponse": { "type": "object", "properties": { "ok": { "type": "boolean" }, "wallet": { "$ref": "#/components/schemas/Wallet" } } },
      "WalletListResponse": { "type": "object", "properties": { "ok": { "type": "boolean" }, "wallets": { "type": "array", "items": { "$ref": "#/components/schemas/Wallet" } } } },
      "UsdAccountRequestResponse": { "type": "object", "properties": { "ok": { "type": "boolean" }, "usdAccountRequest": { "type": "object", "additionalProperties": true, "example": { "reference": "usd_account_request_reference", "status": "pending" } } } },
      "WithdrawRequest": { "type": "object", "required": ["amount", "currency", "bankCode", "accountNumber", "transactionPin"], "properties": { "amount": { "type": "number", "example": 5000 }, "currency": { "$ref": "#/components/schemas/Currency" }, "bankCode": { "type": "string", "example": "000013" }, "accountNumber": { "type": "string", "example": "0123456789" }, "accountName": { "type": "string", "example": "Ada Okafor" }, "description": { "type": "string" }, "transactionPin": { "type": "string", "example": "1234" }, "idempotencyKey": { "type": "string", "description": "Optional body fallback for Idempotency-Key header." } } },
      "CreateCardRequest": { "type": "object", "required": ["walletId", "currency"], "properties": { "walletId": { "type": "string", "format": "uuid" }, "currency": { "$ref": "#/components/schemas/Currency" } } },
      "CardMoneyMovementRequest": { "type": "object", "required": ["amount", "currency", "transactionPin"], "properties": { "amount": { "type": "number", "example": 25 }, "currency": { "$ref": "#/components/schemas/Currency" }, "transactionPin": { "type": "string", "example": "1234" }, "idempotencyKey": { "type": "string" } } },
      "Card": { "type": "object", "properties": { "id": { "type": "string", "format": "uuid" }, "mapleradCardId": { "type": "string", "nullable": true }, "cardLast4": { "type": "string", "nullable": true, "example": "4242" }, "expirationDate": { "type": "string", "nullable": true }, "brand": { "type": "string", "nullable": true }, "currency": { "type": "string" }, "status": { "type": "string", "enum": ["active", "inactive", "blocked"] }, "isFrozen": { "type": "boolean" }, "createdAt": { "type": "string", "format": "date-time" } } },
      "CardResponse": { "type": "object", "properties": { "ok": { "type": "boolean" }, "card": { "$ref": "#/components/schemas/Card" } } },
      "CardMessageResponse": { "type": "object", "properties": { "ok": { "type": "boolean" }, "message": { "type": "string" }, "card": { "$ref": "#/components/schemas/Card" } } },
      "Transaction": { "type": "object", "properties": { "id": { "type": "string", "format": "uuid" }, "userId": { "type": "string", "format": "uuid" }, "senderWalletId": { "type": "string", "format": "uuid", "nullable": true }, "recipientWalletId": { "type": "string", "format": "uuid", "nullable": true }, "type": { "type": "string", "enum": ["deposit", "withdrawal", "transfer"] }, "amount": { "type": "string", "example": "5000.00" }, "currency": { "$ref": "#/components/schemas/Currency" }, "status": { "type": "string" }, "reference": { "type": "string", "nullable": true }, "provider": { "type": "string", "nullable": true }, "providerReference": { "type": "string", "nullable": true }, "providerStatus": { "type": "string", "nullable": true }, "reconciliationStatus": { "type": "string" }, "createdAt": { "type": "string", "format": "date-time" } } },
      "TransactionProviderResponse": { "type": "object", "properties": { "ok": { "type": "boolean" }, "transaction": { "$ref": "#/components/schemas/Transaction" }, "provider": { "type": "object", "properties": { "reference": { "type": "string" }, "status": { "type": "string" } } }, "duplicate": { "type": "boolean" } } },
...
    }
  }
}